resolv add data in the localStorage
