+++ src/components/SearchAndFilter/Query.js
@@ -6,10 +6,12 @@ import FetchAxios from '../../utils/FetchAxios'
 import CardRecipe from '../Card/CardRecipe';
 //import CardWithGrid from '../Card/CardWithGrid';
 // Function
-export default function Query({thingSearched, filter, index, category}) {
+export default function Query({thingSearched, filter, index, category, onChangeRecipe, bla, weekMenu, setWeekMenu}) {
     const [data, setData] = React.useState();
     const [error, setError] = React.useState(null);
-    console.log(`je suis dans query hahaha ${index} et ${category}`)
+    // console.log(`QUERY ==> index :  ${index} et category : ${category}`)
+    // console.log('type de index = ', typeof(index))
+    // console.log('type de category = ', typeof(category))
     React.useEffect(() => {
         const controller = new AbortController();
         if (!thingSearched) {
@@ -41,6 +43,6 @@ export default function Query({thingSearched, filter, index, category}) {
         )
     }
     return (
-        <CardRecipe index={index} category={category} datas={data}/>
+        <CardRecipe indexOfDay={index} categoryOfMeal={category} onChangeRecipe={onChangeRecipe} bla={bla} weekMenu={weekMenu} setWeekMenu={setWeekMenu} datas={data}/>
     )
 }
