--- src/components/data/recipeFake.js
@@ -1,17 +1,0 @@
-'./fraise.jpg'
-export const meal = [
-    {id : "R1", titre : "Vienoiserie", category :"matin",picture : '../images/fraise.jpg', recipe:"text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic"},
-    {id : "R2", titre : "Céréales au lait", category :"matin", picture :'../images/fraise.jpg', recipe:"text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic"},
-    {id : "R3", titre : "Fruit", category :"matin", picture :'../images/fraise.jpg', recipe:"text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic"},
-    {id : "R4", titre : "Patate farci", category :"diner", picture :'../images/snake.jpg', recipe:"text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic"},
-    {id : "R5", titre : "Boeuf fou au carotte", category :"diner", picture :'../images/snake.jpg', recipe:"text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic"},
-    {id : "R6", titre : "Boeuf fou au carotte", category :"diner", picture :'../images/snake.jpg', recipe:"text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic"},
-    {id : "R7", titre : "Poissons qui perd ses os", category :"diner", picture :'../images/snake.jpg', recipe:"text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic"},
-    {id : "R8", titre : "Rizoto aux champignion", category :"diner", picture :'../images/snake.jpg', recipe:"text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic"},
-    {id : "R9", titre : "Ragout de murloc", category :"diner", picture :'../images/snake.jpg', recipe:"text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic"},
-    {id : "R10", titre : "Cuissot de worg et ses ichor", category :"deujeuner", picture :'../images/salad.jpg', recipe:"text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic"},
-    {id : "R0", titre : "Galette de Courgette rapé", category :"deujeuner",picture :'../images/salad.jpg', recipe:"text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic"},
-    {id : "R10", titre : "Lasagne bolognaise", category :"deujeuner",picture :'../images/salad.jpg', recipe:"text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic"},
-    {id : "R11", titre : "Salade césar", category :"deujeuner", picture :'../images/salad.jpg', recipe:"text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic"}
-]
-
