--- src/components/data/weekMenu.js
@@ -1,47 +1,0 @@
-export let weekMenu2 =[ 
-    {day : "lundi" , matin : "", deujeuner : "", diner : ""},
-    {day : "mardi" , matin : "", deujeuner : "", diner : ""},
-    {day : "mercredi" , matin : "", deujeuner : "", diner : ""},
-    {day : "jeudi" , matin : "", deujeuner : "", diner : ""},
-    {day : "vendredi" , matin : "", deujeuner : "", diner : ""},
-    {day : "samedi" , matin : "", deujeuner : "", diner : ""},
-    {day : "dimanche" , matin : "", deujeuner : "", diner : ""}
-    
-]
-
-export let weekMenu =[ 
-    {lundi : {matin : "patantouille",deujeuner : "",diner : ""}},
-    {mardi :{
-        matin : "",
-        deujeuner : "",
-        diner : ""}},
-    {mercredi :{
-        matin : "",
-        deujeuner : "",
-        diner : ""
-    }},
-    {jeudi :{
-        matin : "",
-        deujeuner : "",
-        diner : ""
-    }},
-    {vendredi : {
-        matin : "",
-        deujeuner : "",
-        diner : ""
-    } },
-    {samedi :{
-        matin : "",
-        deujeuner : "",
-        diner : ""
-    }},
-    {dimanche : {
-        matin : "",
-        deujeuner : "",
-        diner : ""
-    }}
-
-]
-
-
-
