+++ src/userApp.js
@@ -9,9 +9,8 @@ import Home from './pages/Home'
 import MyAccount from './pages/MyAccount';
 import { Box } from '@chakra-ui/react'
 import SearchAndFilter from './components/SearchAndFilter/SearchAndFilter';
-import CreateMenu from './components/meal/createMenu';
 import DayMeal from './components/meal/dayMeal';
-import Mealy from './components/mealy/mealy';
+import MenuCreator from './components/menuCreator/menuCreator';
 
 const UserApp = () => {
     return (
@@ -24,10 +23,9 @@ const UserApp = () => {
                 <Route path='/login' element={<Login />} />
                 {/* <Route path='/user' element={<NutriHomeUser />} /> */}
                 <Route path='/findRecipe' element={<SearchAndFilter />} />
-                <Route path='/createmenu' element={<CreateMenu />} />
                 <Route path='/menu' element={<DayMeal/>} />
                 <Route path='/myaccount' element={<MyAccount />} />
-                <Route path='/mealy' element={<Mealy />} />
+                <Route path='/ronan' element={<MenuCreator />} />
                 {/*<Route path='/menu' element={<UserMenu/>} />*/}
                 {/*<Route path='/menu/:id' element={<MenuDetail/>} />*/}
                 {/*<Route path='/recipe/:id' element={<Recipe/>} />*/}
